Disable scrolling to table row as this was crashing
This is a change to a third-party library that appears abandoned.
This fix is a hack. There is likely a proper way to fix this while
maintaining the scrolling behavior, but this fixes the main crashes.
The app is slated for a rewrite, so this will likely do for now.